requirements.txt
Django==1.11.4
django_jsonfield==1.0.1
